Remove dead software from sucks.md
When projects haven't received care for over 7 years, they should
make room for more recent examples of suckness.
There's no reason to encourage bit-rot.
The first step is to clean this list, before putting in newer
elements.
While at it, shuffle around a bit more with other areas of this
page to concentrate the message.