Vectorize dwm-logo
png is a great image format, especially for line-art. However,
in this case, going with svg brings some benefits:
Zooming into png-images, you get artifacts from the png-renderer.
This can easily be reproduced by zooming into the page and taking a
close look at the edges.
This means that upscaling a png in situ is not trivial. Also, the
source-code is now human-readable.