Remove Lwan web server from rocks.md
Here are the reasons:
   - It depends on cmake (which _sucks_)
   - It depends on Python (no need for that for a web server)
The build system pulls in other stuff if available, like Lua, SQLite,
MySQL. So it's a mixed bag of a lot of cruft.

The server may be fast as hell, but it definitely comes at the cost of
simplicity. And suckless.org is not primarily about speed, but
simplicity and correctness.
The codebase is not inviting to start working on it. Different malloc-
implementations complicate the readability of the code even more.

And of course, the rocks.md-page is not an advertising page. Don't
misuse it to get some in-links to your new web server website,
especially if you are not really adhering to the suckless philosophy
or even refrain from using tools listed in sucks.md.