dwm swallow patch: update patch to make it easier to use
I seriously spent 1 hour figuring out why the patch didn't work
until I found out you had to do a -DSWALLOWING in config.mk and
add a bunch of libs for xcb to work.
I didn't like that.

Thus, I updated the patches to actually work out of the box without
shitting on the user. I also reworked the patch page so you can
see what the hell is going on with this shit.

It's a very cool patch and I love it in everyday work!