Update invert.c according to repo-changes
It was my interest to create an example that is not just an
artificial demonstration, but that can actually be used in
"production". This bloats the code up a bit, but with the comments
and so on it actually gets much more readable.

The 0BSD allows use without any strings attached.