This commit was suggested to me by Anselm Garbe.
The index.md is edited to list the submitted diff-file and my
email contact.

The st-alpha-0.8.1.diff is basically a modified st-alpha-20171221-0ac685f.diff
with mostly line changes to where the hunks are supposed to be placed.
The hunk that originally went to 'win.h' is now targeted at 'x.c'

Pleas rech back to me, if there would be anything to change since this
is actually my fisrt code commit to a git.