sucks: re-add GNU make
GNU make is ubiquitous, even the lame BSDs have it installed and there's no
reason to not use its advanced features over the dumb-ass POSIX make.
It's still slim and highly portable. It builds in 2 seconds from source on any
half-ways modern system.
from all the modern build environments it's the best choice, and that's the
reason why musl uses it.