Static website generation
This is a first proposal of a trivial, ad-hoc static website generatior
for the suckless.org website.

It uses smu for markdown parsing, so some pages may need editing.

A static website can be generated with ./make
The files in _werc are no longer needed.
The files title, */title, domains, head.html, foot.html, links.html
are added for configuration.