./make: add missing directory listing
The previous verision did not handle directories without a index.md
file in.  This version recurse over each directory instead of each
markdown file so that even empty directories get an "ls page".

Also, a scheme is added in front of the blue nav bar links so that
they do not get treated as relative links but rather absolute links.