[dmenu][patch][fuzzymatch] Port to 4.9
Add dmenu-fuzzymatch-4.9.diff because dmenu-fuzzymatch-4.6.diff produces a warning during the compilation of dmenu. dmenu-fuzzymatch-20170603-f428f3e.diff fixes this and adds an option to config.def.h and the flaf -F to dmenu which enable to turn fuzzymatching on and off. However, dmenu-fuzzymatch-20170603-f428f3e.diff can not be applied to dmenu-4.9 without error. dmenu-fuzzymatch-4.9.diff fixes the error.