[dwm][patches][nostatusbar] Delete patch
I created this patch and now I am deleting it because I don't think it
provides any value. Somebody who does not want to use the statusbar can
hide it. Hence it does not add any new functionality or improves the
workflow. On the other hand it can make things more difficult because it
removes a lot of code and applying new patches results often in conflicts.
However, I still think that the statusbar should not be hardcoded in dwm
but rather be a separate program.