[surf][patch] externalpipe: Export full document outerHTML when JS is enabled
Rather than exporting only the body when JS is enabled; now the non-JS
functionality is matched and we export the entire HTML document. This fixes
a bug where symbols / non-english characters are not rendered properly because
the meta tag is not recognized in the associated surf_linkselect.sh script's
xmllint call.