[dwm][focusonnetactive] Minor multi-monitor fix for the focusonnetactive patch
In a multi-monitor setup you can have an application starting up on tag 8 on monitor 1. With the original patch if the selected monitor was monitor 2 at the time, then you'd view tag 8 on monitor 2 rather than monitor 1 where the application started.

This is a minor fix to change the selected monitor to that of the client before viewing the tag.