Prevent reset of SIGCHILD handler to default
This patch fixes a bug in the externalpipe patch, and should be applied
on top of it. It prevents the reset of the signal handler set on
SIGCHILD, when the forked process that executes the external process
exits. I opted for switching from signal to sigaction instead of
rearming the signal in the sigchld function, just because it is the
recommended function (although I tried both ways and both worked).