Implements TODO to update borders on bg color change
When an OCS sequence was used to change the bg color, the borders where
dirty. This simple patch just clears the window before the redraw of the
terminal when the bg color has been changed. This is apparently smooth
and solves the problem. There was a TODO comment for it on the st.c
file, which I removed.