Added 6.2 patch for moveresize.
The older patches did not include default keybindings and the examples
listed on the page are incompatible with any of the patches listed.
They are only for the inline function listed on the same page.

The later versions of the moveresize patch also support setting
absolute position and/or size.

Adding this primarily to avoid having other people waste time working
out what the keybindings should look like.