Revert "[st][patch][lowlatency] Introducing new patch for st called lowlatency."
This reverts commit 87fac66fea613a7e8d798da65dcd3552d4766817.

Assuming the author probably had good intentions, but the page contains too
much inaccurate and wrong information so it is reverted.

st has had work on improving the latency and drawing performance (less flicker):
https://git.suckless.org/st/commit/1d590910652519268152eae6b97cf30ace4e90c0.html

There you can lower the minlatency and maxlatency values. This is a better
solution than forcing drawing after each keypress as done in this patch
st.suckless.org/patches/lowlatency/st-lowlatency-0.8.3.diff
Which causes flicker and potentially reduce throughput aswell (as mentioned in
the notes section). The auto-sync patch (in master) does not have this issue.
Before the auto-sync patch it was already possible to increase the xfps and
actionfps causing lower latencies aswell (at some other costs potentially).

st was never as slow as the benchmark referenced on the page:
https://danluu.com/term-latency/
There the benchmarks are run with a Java application on MacOS:
"Measurements are with macOS unless otherwise stated."
This is not a reliable way to test. The proprietary MacOS is not supported
either.

For an accurate benchmark provide _atleast_:

* A listing of exactly all the versions used.
* System information.
* An FOSS OS (Linux, BSD).
* A reliable simple reproducable benchmarking tool (open-source of course).
* The exact dataset used for testing.