Point to bestmotherfucking.website (without leading `the')
This reverts commit 757f08c89ccca636c84725fd60a9fbd1ad1250e5.

The commit said that it fixed the link but there was nothing to fix, i.e. the
domain and web site was still valid and the one with leading `the' seems just a
different version.

That version seems to have all the HTML code entirely in one line and also
inject tracking via javascript.

The main point of these kind of websites are to get some fun and by adding
tracking, javascript or similar the joke is getting a bit sad.

(I have tried to contact the original committer about that pointing out such
problems but I have not any reply.)