[dwm][patch] fullgaps-toggle
Modified version of fullgaps

[Alt]+[Shift]+[=] toggles gaps. Re-enabling gaps will return to the
previous gap size.

[Alt]+[Shift]+[-] resets gaps to default
value in `conifg.h`.

The Gap struct is defined as follows:

typdef struct {
	int isgap;
	int realgap;
	int gappx;
} Gap;

See `void setgaps(Arg *arg) {...}` for how this works.

(This is the same as commit 3b8e1d672474f842ff111 but with a proper
email address. Please ignore that commit.)