[dmenu][patch][highlight] Add revised patch
This commit adds a minorly revised version of the original highlight
patch by Miles Alan. This revision introduces a new array element in
*colors[SchemeLast][2], called [SchemeOutHighlight], which sets the
foreground and background colours for matched text in multiple selection
mode (previously missing in the original patch), along with necessary
changes to dmenu.c. This patch has been tested and applies cleanly to
the master branch of dmenu.