Feat: Status monitor akuma-v-dwm
Add akuma-v-dwm status monitor to the "Status Monitors Submitted by
Others" section.

Happy New Year!

I've been using dwm for a little while now, and I really like it,
as well as the suckless philosophy, so much so that I was inspired to
write this program. It is not in C, but a shell script (I'm thinking a
C version will be my next project). I believe it is a little different
than the other shell status monitors already on the site, in that
it is event driven and only recomputes the parts of the status that
are requested.

Thanks for all of your hard work, and good luck on your endeavours
in the new year!