Added dwm default-tag-apps patch
This is a new patch i've wanted to add for a while and one i've been
using on my own. I have tried to follow all the other patch layouts and
instructions, but in my own smu renders of the index.md file the email
does not show. I figure however that it is better to leave the file in
the same format as the others so that's what i've done.