[dwm][patch][statuscmd] Add various tweaks, fixes and updates
The signal version is now the default and the dwmblocks patch is now up
to date and forks the commands it runs.

The signal-less version is now simpler to use because it now uses an
array of command and number pairs instead of relying on the correct
index being specified.

There are also compatibility patches for status2d now.