[dwm][patch][onlyquitonempty] add updated version to fix bugs
(Dates are a bit weird: patch committed to my dwm repo June 2020, patch
apparently made December 2020, committed here April 2021)

This is the new version of onlyquitonempty, which fixes a few bugs
compared to the old one. It also has new documentation, to make it
easier to use.

The primary bugfix is that a temporary allocation 'junk' has been
removed. This would have been overflowed in any case, and has been
replaced with stack variables. Unfortunately, X demands to give us
'children', so we need to XFree that, but oh well. (Potentially that's
a memory leak fixed?)

In addition, an override has been added so you can still get out if you
mess up your configuration.