tools.suckless.org: remove zsh-ism in example.
`do' requires at least one command after it in POSIX sh and bash.
`do;' is invalid in most shells, but zsh.
The `;' was unnecessary so I just removed it.

I also quoted a command substitution that was not quoted to avoid
problems in case `xdotool' is not installed.

And I also removed/added some whitespace to clean up the file.