[dwm][patches][barpadding] fix window area misalignment for adjacent monitors
This fixes a bug with the original patch where the sp and vp variables
were set after the call to updategeom() in the setup function. This
causes the window area to not take the vertical spacing into account and
it will be misaligned on startup.

This was worked around by adding another call to updatebarpos(selmon)
which works for the primary monitor, but leaves the misalignment in place
for any adjacent monitor in a multi-monitor setup.

The fix for this is simply to move the setting of the sp and vp variables
before the updategeom() call. Subsequently the call to updatebarpos can
be removed as well.