[dmenu][separator]: rebase + fix invalid free()
when `separator_reverse` is active `items.text_output` and `items.text`
will be swapped leading to trying to free a pointer that wasn't
allocated.

this was not a problem before b43ec05 since `items` were not being
free-ed at cleanup() before that commit.

also rm `separator_greedy` and use a function pointer instead.