[dmenu][mouse-support]: fix bug and update description
Looks like I unknowningly reinvented the "hover-bg" version of the
patch. I hadn't noticed it since it was at the bottom of the list where
outdated patches typically are.

Looking at the "hover-bg" variant revealed a bug in the new patch where
promptw wasn't being accounted for.

But in any case, doesn't seem like there's any reason to have two
separate versions. If there are some reason for the separation then feel
free to pull the motionevent() into a separate diff.