[other_projects]: move nt in here from rocks
IMO, rocks page should contain software that's proven to be useful in
practice, and endorsed by people on the "suckless community". it should
not be a page for "advertisement".

the `nt` project doesn't seem to have much users, and it was added to
the rocks page by the author himself. and so move it to the
other_projects page, where it's more appropriate.