rocks: add senpai with soju
senpai is a simple Go IRC client with only 8 Go module dependencies,
and when used alongside soju, is a very comfortable experience. at the
moment, there is no simple bouncer for IRC, other than soju and pounce,
which are both very great when used with their appropiate clients. i've
taken a look at irssi, and while it isn't suckless, i can understand why
people swear by it, since it is very extensible - perhaps the reason why
it uses glib.

aerc is also a very nice and comfortable mail client with 50 Go modules,
and is kept on the list, since it is a very nice solution.