rocks: remove irssi - NO REASON TO KEEP OTHER THAN BIAS
irssi uses glib.
irssi needs OpenSSL, and Perl.
irssi needs Meson.
irssi is 60 thousand lines of C.
This is against the suckless philosphy.
suckless has a strong opinion on glib: https://suckless.org/sucks/

If this commit is reverted - then rocks is biased and no longer supports
software that is on-par with the suckless philosphy, and cannot
be trusted.