[dwm][patch][betterswallow] Add patch
This patch is yet another take on swallowing, this time with a PID-based
mechanism for detecting swallowed windows and the simple IPC mechanism of
a ClientMessage.

The PID-based mechanism is, in my opinion, significantly better than the
crude string matching mechanisms of `dynamicswallow`, not to mention the
`devour` program which just unmaps the active window.

Since IPC is handled with a ClientMessage, the code for IPC is
significantly simpler than alternative approaches and is entirely
self-contained in this patch.